Accomodate for groff's crappy behaviour wherein an unrecognised single-character escape (and ONLY this type of escape) will map back into itself:
       "If a backslash is followed by a character that does not
	constitute a defined escape sequence the backslash is silently
        ignored and the  character maps to itself."

(From groff.7.)

Found by Jason McIntyre.

+++ mandoc.c
@@ -314,7 +314,7 @@ mandoc_eos(const char *p, size_t sz, int enclosed)
 	 */
 
 	found = 0;
-	for (q = p + sz - 1; q >= p; q--) {
+	for (q = p + (int)sz - 1; q >= p; q--) {
 		switch (*q) {
 		case ('\"'):
 			/* FALLTHROUGH */
